Dealer 타입 추가
