+++ src/main/kotlin/blackjack/Dealer.kt
@@ -1,0 +1,8 @@
+package blackjack
+
+class Dealer() : User(NAME) {
+
+    companion object {
+        const val NAME = "딜러"
+    }
+}
