+++ src/main/kotlin/blackjack/Player.kt
@@ -1,22 +1,6 @@
 package blackjack
 
-import blackjack.Card.Companion.BLACK_JACK_NUM
-
-class Player(val name: String) {
-    val cardDeck = Cards()
-
-    init {
-        require(name.isNotBlank()) { "이름이 공백이어서는 안됩니다." }
-    }
-
-    fun addCard(card: Card) {
-        cardDeck.add(card)
-    }
-
-    fun isDead(): Boolean {
-        return cardDeck.getScore() > BLACK_JACK_NUM
-    }
-}
+class Player(name: String) : User(name)
 
 class Players(private val _players: List<Player>) {
     val players: List<Player>
